move SpatialProfile and SpectralDistribution to core
