+++ bundle/edu.gemini.itc.shared/src/main/scala/edu/gemini/itc/shared/Parameters.scala
@@ -1,8 +1,7 @@
 package edu.gemini.itc.shared
 
-import edu.gemini.spModel.core.{Redshift, BrightnessUnit, MagnitudeBand}
+import edu.gemini.spModel.core.{SpectralDistribution, UniformSource, SpatialProfile, Redshift, BrightnessUnit, MagnitudeBand}
 import edu.gemini.spModel.gemini.obscomp.SPSiteQuality.{WaterVapor, CloudCover, SkyBackground, ImageQuality}
-import edu.gemini.spModel.target.{UniformSource, SpectralDistribution, SpatialProfile}
 
 // ==== Observing conditions
