+++ bundle/edu.gemini.itc.web/src/main/java/edu/gemini/itc/web/html/Flamingos2Printer.java
@@ -7,8 +7,8 @@ import edu.gemini.itc.flamingos2.Flamingos2;
 import edu.gemini.itc.flamingos2.Flamingos2Recipe;
 import edu.gemini.itc.shared.*;
 import edu.gemini.spModel.gemini.flamingos2.Flamingos2.FPUnit;
-import edu.gemini.spModel.target.PointSource$;
-import edu.gemini.spModel.target.UniformSource$;
+import edu.gemini.spModel.core.PointSource$;
+import edu.gemini.spModel.core.UniformSource$;
 import scala.Tuple2;
 import scala.collection.JavaConversions;
