+++ bundle/edu.gemini.itc.web/src/main/scala/edu/gemini/itc/web/html/HtmlUtil.scala
@@ -1,7 +1,7 @@
 package edu.gemini.itc.web.html
 
 import edu.gemini.itc.shared.SourceDefinition
-import edu.gemini.spModel.target._
+import edu.gemini.spModel.core._
 
 /**
  * HTML Utils
