+++ bundle/edu.gemini.spModel.core/src/main/scala/edu/gemini/spModel/core/SpatialProfile.scala
@@ -1,4 +1,4 @@
-package edu.gemini.spModel.target
+package edu.gemini.spModel.core
 
 sealed trait SpatialProfile extends Serializable
 case object PointSource extends SpatialProfile
